#10612: Moved unit tests to separate file.
